Move header, footer to common
