+++ src/Components/Common/Footer/Footer.tsx
@@ -1,5 +1,5 @@
 import * as S from "./styled";
-import * as I from "../../Asset/SVG/index";
+import * as I from "../../../Asset/SVG/index";
 
 const Footer: React.FC = () => {
   return (
