+++ src/Components/Common/Header/Header.tsx
@@ -1,5 +1,5 @@
 import * as S from "./styled";
-import * as I from "../../Asset/SVG/index";
+import * as I from "../../../Asset/SVG/index";
 import { Link } from "react-router-dom";
 
 interface HeaderProps {
